Add files via upload
Used several algorithms like RandomForest, DecisionTree, AdaBoost, GradientBoosting, SVM with different kernels. Removed a few other columns with large number of values (more than 99%) belonging to a single category. Also performed onehotencoding and normalized the data. Despite everything the highest accuracy for the part of the training set is 0.57, I think we have to check it once on the whole dataset.